move readproxy
