+++ ydb/core/persqueue/pqtablet/common/constants.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <util/datetime/base.h>
 #include <util/generic/size_literals.h>
 #include <ydb/core/persqueue/public/constants.h>
 
@@ -12,5 +13,7 @@ static const ui32 MAX_BLOB_PART_SIZE = 500_KB;
 //
 // Total time till kafka supportive partition deletion = AppData.KafkaProxyConfig.TransactionTimeoutMs + KAFKA_TRANSACTION_DELETE_DELAY_MS
 static const ui32 KAFKA_TRANSACTION_DELETE_DELAY_MS = TDuration::Hours(1).MilliSeconds(); // 1 hour;
+
+static constexpr char TMP_REQUEST_MARKER[] = "__TMP__REQUEST__MARKER__";
 
 }
