+++ ydb/core/persqueue/pqtablet/readproxy/ya.make
@@ -1,0 +1,17 @@
+LIBRARY()
+
+SRCS(
+    readproxy.cpp
+)
+
+
+
+PEERDIR(
+    ydb/core/persqueue/events
+    ydb/core/persqueue/common
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+)
